+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed bug where timer would go one last cycle after logging out.
 * Added some navigation improvments.
 * Avatar changes now go through the api (and actually change. wow!)
+* Added record feature to record view controller (doesn't contact api yet.)
 ## v0.2 (PRERELEASE)
 * Added more error handling
 * Some more constraints